Move constants check to its own package
